Reorganized page URL helpers
